refactorización de codigo
